requirements.txt
fastapi==0.115.8
nltk==3.8.1
pandas==1.3.5
pydantic==2.10.6
python-dotenv==1.0.1
transformers==4.30.2